pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "web-services-infrastructure"
version = "1.0.0"
description = "Multi-client web development services infrastructure using AWS CDK"
authors = [
    {name = "Web Services Infrastructure Team"}
]
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    # AWS CDK Core - minimal for Week 1
    "aws-cdk-lib>=2.100.0",
    "constructs>=10.0.0",
    # Basic utilities for client config validation
    "pydantic>=2.0.0",
    "boto3>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "aiohttp>=3.13.0",
]

[tool.black]
line-length = 88
target-version = ['py313']

[tool.ruff]
target-version = "py313"
line-length = 88
select = ["E", "W", "F", "I"]

[tool.hatch.build.targets.wheel]
packages = ["stacks", "clients", "shared", "tools", "constructs"]
